# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Vida Invoice API
  version: 1.0.0
  description: >
    API for creating UBL invoices from e-commerce order webhooks and retrieving generated artefacts.
servers:
  - url: http://localhost:3001
    description: Local development
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
    InvoiceResponse:
      type: object
      required:
        - invoiceId
      properties:
        invoiceId:
          type: string
          description: Identifier of the generated invoice.
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        details:
          description: Additional error details.
        code:
          type: string
        ruleId:
          type: string
    HistoryEntry:
      type: object
      required:
        - requestId
        - timestamp
        - status
        - durationMs
      properties:
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
        source:
          type: string
        orderNumber:
          type: string
        originalOrderId:
          type: string
        tenantId:
          type: string
        status:
          type: string
          enum: [ok, error]
        invoiceId:
          type: string
        invoicePath:
          type: string
        durationMs:
          type: integer
        error:
          type: string
        peppolStatus:
          type: string
        peppolId:
          type: string
paths:
  /webhook/order-created:
    post:
      summary: Create an invoice from a storefront order webhook
      description: >
        Normalises incoming order payloads, generates a UBL invoice, persists it to disk,
        and optionally dispatches it to a PEPPOL Access Point.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - source
                - payload
                - supplier
              properties:
                source:
                  type: string
                  enum: [shopify, woocommerce, order]
                payload:
                  description: Source order payload.
                supplier:
                  description: Supplier party information.
                defaultVatRate:
                  type: number
                  enum: [0, 6, 12, 21]
                currencyMinorUnit:
                  type: integer
                  minimum: 0
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema:
            type: string
          description: >
            Optional key to ensure duplicate deliveries receive the same invoice identifier.
      responses:
        "200":
          description: Invoice created.
          headers:
            X-Idempotency-Cache:
              schema:
                type: string
              description: Indicates whether the response came from the idempotency cache.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
        "400":
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Authentication failed.
        "422":
          description: UBL validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "429":
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "502":
          description: Access Point send failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /invoice/{invoiceId}:
    get:
      summary: Retrieve a generated UBL invoice
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: invoiceId
          required: true
          schema:
            type: string
          description: Identifier returned from the webhook endpoint.
      responses:
        "200":
          description: UBL XML invoice.
          content:
            application/xml:
              schema:
                type: string
        "400":
          description: Invalid identifier supplied.
        "401":
          description: Authentication failed.
        "404":
          description: Invoice not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Unexpected server error.
  /history:
    get:
      summary: List recent webhook executions
      security:
        - ApiKeyAuth: []
      parameters:
        - in: query
          name: tenant
          schema:
            type: string
          description: Filter events for a specific tenant identifier.
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 200
          description: Maximum number of history entries to return.
      responses:
        "200":
          description: Recent webhook executions.
          content:
            application/json:
              schema:
                type: object
                properties:
                  history:
                    type: array
                    items:
                      $ref: '#/components/schemas/HistoryEntry'
        "401":
          description: Authentication failed.
        "500":
          description: Unexpected server error.
  /_health:
    get:
      summary: Health check
      description: Returns HTTP 200 when the service is ready to serve traffic.
      responses:
        "200":
          description: Service is healthy.
          content:
            text/plain:
              schema:
                type: string
  /metrics:
    get:
      summary: Prometheus metrics
      description: Exposes application counters in Prometheus exposition format.
      responses:
        "200":
          description: Prometheus metrics payload.
          content:
            text/plain:
              schema:
                type: string